requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
requests==2.31.0
reportlab==4.0.7
pillow==10.1.0
qrcode==7.4.2
pyhanko==0.20.1
cryptography==41.0.7
openpyxl==3.1.2
pikepdf==8.7.1
pytz==2024.1
redis==5.0.1
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1
psutil==5.9.6

